Update function names and adding sub modules for WorkspaceMember and WorkflowProcess
