+++ api/job.py
@@ -3,6 +3,7 @@ from urlparse import urljoin
 from urlparse import urlparse
 from api.exception import *
 from api.alpineobject import AlpineObject
+from api.jobtask import JobTask
 
 
 class Job(AlpineObject):
@@ -10,6 +11,7 @@ class Job(AlpineObject):
     Setup and run jobs and tasks
     """
 
+    task = None
     def __init__(self, base_url, session, token):
         super(Job, self).__init__(base_url, session, token)
         self.chorus_domain = '{uri.scheme}://{uri.netloc}/'.format(uri=urlparse(self.base_url))
@@ -17,37 +19,7 @@ class Job(AlpineObject):
         self.alpine_base_url = urljoin(self.chorus_domain,
                                        "alpinedatalabs/api/{0}/json".format(self._alpine_api_version))
         self.logger.debug("alpine_base_url is: {0}".format(self.alpine_base_url))
-
-    def run(self, job_id):
-        """
-        Run a job.
-
-        :param string workflow_id:
-        :param string variables:
-        :return:
-        :rtype: str
-        """
-
-        url = "{0}/jobs/{1}/run?saveResult=true".format(self.base_url, job_id)
-        print(url)
-
-        self.session.headers.update({"x-token": self.token})
-        self.session.headers.update({"Content-Type": "application/json"})
-
-        response = self.session.post(url, timeout=30)
-
-        return response
-
-        # self.session.headers.pop("Content-Type")
-        # self.logger.debug(response.content)
-        # if response.status_code == 200:
-        #     process_id = response.json()['meta']['processId']
-        #
-        #     self.logger.debug("Workflow {0} started with process {1}".format(workflow_id, process_id))
-        #     return process_id
-        # else:
-        #     raise RunFlowFailureException(
-        #         "Run Workflow {0} failed with status code {1}".format(workflow_id, response.status_code))
+        self.task = JobTask(base_url, session, token)
 
     def create(self, workspace_id, job_name, interval_unit="on_demand", interval_value=0, next_run=""):
         """
@@ -80,7 +52,25 @@ class Job(AlpineObject):
         self.logger.debug("Received response code {0} with reason {1}...".format(response.status_code, response.reason))
         return response.json()['response']
 
-    def get_all(self, workspace_id, per_page=50):
+    def delete(self, workspace_id, job_id):
+        """
+        Delete job from a workspace
+
+        :param workspace_id: Id of the workspace to be deleted in
+        :param job_name: name of the job to delete
+        :return: response for the delete action
+        """
+
+        url = "{0}/workspaces/{1}/jobs/{2}".format(self.base_url, workspace_id, job_id)
+        url = self._add_token_to_url(url)
+
+        # POSTing a HTTP delete
+        self.logger.debug("Deleting the job id: <{0}> from workspace id: <{1}>".format(job_id, workspace_id))
+        response = self.session.delete(url, verify=False)
+        self.logger.debug("Received response code {0} with reason {1}".format(response.status_code, response.reason))
+        return response
+
+    def get_list(self, workspace_id, per_page=50):
         """
         Get a list of jobs from a workspace
 
@@ -108,7 +98,7 @@ class Job(AlpineObject):
                 break
         return jobs_list
 
-    def get_data(self, workspace_id, job_id):
+    def get(self, workspace_id, job_id):
         """
         Get Job info from a workspace
 
@@ -142,149 +132,40 @@ class Job(AlpineObject):
         :param job_name: name of the job to query on
         :return:
         """
-        job_list = self.get_all(workspace_id)
+        job_list = self.get_list(workspace_id)
         for job_info in job_list:
             if job_info['name'] == job_name:
                 return job_info['id']
         # return None
         raise JobNotFoundException("Job {0} not found".format(job_name))
 
-    def delete(self, workspace_id, job_id):
+    def run(self, job_id):
         """
-        Delete job from a workspace
+        Run a job.
 
-        :param workspace_id: Id of the workspace to be deleted in
-        :param job_name: name of the job to delete
-        :return: response for the delete action
-        """
-
-        url = "{0}/workspaces/{1}/jobs/{2}".format(self.base_url, workspace_id, job_id)
-        url = self._add_token_to_url(url)
-
-        # POSTing a HTTP delete
-        self.logger.debug("Deleting the job id: <{0}> from workspace id: <{1}>".format(job_id, workspace_id))
-        response = self.session.delete(url, verify=False)
-        self.logger.debug("Received response code {0} with reason {1}".format(response.status_code, response.reason))
-        return response
-
-    def add_task(self, workspace_id, job_id, workfile_id, task_type="run_work_flow"):
-        """
-        Add a task to a job
-
-        :param workspace_id: Id of the workspace for the job
-        :param job_id: Name of the job for which the task is to be added
-        :param workfile_id: Id of the workfile to be added as a task
-        :param task_type: task type could be run_work_flow or run_sql_workfile
-        :return: Info of the new added task
-
-        """
-        self.logger.debug("The job id of the job id: <{0}>".format(job_id))
-        url = "{0}/workspaces/{1}/jobs/{2}/job_tasks".format(self.base_url, workspace_id, job_id)
-        url = self._add_token_to_url(url)
-        self.logger.debug("The URL that we will be posting is: {0}".format(url))
-
-        # constructing the payload for adding a task
-        payload = { "action": task_type, "workfile_id": workfile_id}
-
-        self.logger.debug("POSTing payload {0} to URL {1}".format(payload, url))
-        response = self.session.post(url, data=payload, verify=False)
-        self.logger.debug("Received response code {0} with reason {1}...".format(response.status_code, response.reason))
-        return response.json()['response']
-
-    def add_workflow_task(self, workspace_id, job_id, workflow_id):
-        """
-        Add a workflow task to a job
-
-        :param workspace_id: Id of the workspace for the job
-        :param job_id: Name of the job for which the task is to be added
-        :param workflow_id: Id of the workflow to be added as a task
-        :return: Info of the new added task
-        """
-        # Setting the task type to run_work_flow, since this is adding a workflow as a task to a job on a workspace
-        task_type = "run_work_flow"
-        return self.add_task(workspace_id, job_id, workflow_id, task_type)
-
-    def add_sqlworkfile_task(self, workspace_id, job_name, sql_workfile_id):
-        """
-        Add a Sql Workfile task to a job
-
-        :param workspace_id: Id of the workspace for the job
-        :param job_name: Name of the job for which the task is to be added
-        :param sql_workfile_id: Id of the sql workfile to be added as a task
-        :return: Information of the new added task
-        """
-        sql_task_type = "run_sql_workfile"
-        return self.add_task(workspace_id, job_name, sql_workfile_id, sql_task_type)
-
-    def get_tasks(self, workspace_id, job_id):
-        """
-
-        :param workspace_id:
-        :param job_name:
+        :param string workflow_id:
+        :param string variables:
         :return:
+        :rtype: str
         """
 
-        self.logger.debug("Getting the Job id of Job: {0}".format(job_id))
-        self.logger.debug("Retrieved the Job id of the job: {0} to be: {1}".format(job_id, job_id))
+        url = "{0}/jobs/{1}/run?saveResult=true".format(self.base_url, job_id)
+        print(url)
 
-        # Constructing the URL to retrieve the contents
-        url = "{0}/workspaces/{1}/jobs/{2}".format(self.base_url, workspace_id, job_id)
-        url = self._add_token_to_url(url)
+        self.session.headers.update({"x-token": self.token})
+        self.session.headers.update({"Content-Type": "application/json"})
 
-        # Doing a HTTP GET
-        self.logger.debug("Posting a HTTP GET to retrieve the tasks on the workspace.")
-        response = self.session.get(url)
-        self.logger.debug("Received response code {0} with reason {1}...".format(response.status_code, response.reason))
-        return response.json()['response']['tasks']
+        response = self.session.post(url, timeout=30)
 
-    def get_task_data(self, workspace_id, job_id, task_id):
-        """
-        Get the infomation of a task
-
-        :param workspace_id: Id of the workspace for the job
-        :param job_name: Name of the job
-        :param task_name: Name of the Task
-        :return: Information of the task
-        """
-        task_list = self.get_tasks(workspace_id, job_id)
-        for task in task_list:
-            if task['id'] == task_id:
-                self.logger.debug("We have successfully verified that we have created the task id: <{0}>".format(task_id))
-                return task
-        raise TaskNotFoundException("The Task with id: <{0}> doesn't exits".format(task_id))
-
-    def get_task_id(self, workspace_id, job_id, task_name):
-        """
-        Get the Id of a task in a job
-
-        :param workspace_id: Id of the workspace for the job
-        :param job_name: Name of the job
-        :param task_name: Name of the task
-        :return: Id of the task
-        """
-        task_list = self.get_tasks(workspace_id, job_id)
-        for task in task_list:
-            if task['name'] == task_name:
-                self.logger.debug(
-                    "We have successfully verified that we have created the task: {0}".format(task_name))
-                return int(task['id'])
-        return None
-        # raise TaskNotFoundException("The Task with name: {0} doesn't exits".format(task_name))
-
-    def delete_task(self, workspace_id, job_id, task_id):
-        """
-        Delete a task from the job on a workspace
-
-        :param workspace_id: Id of the workspace from which the task has to be deleted
-        :param job_name: Name of the job to from which the task is to be deleted
-        :param task_name: Name of the task
-        :return: Response of the delete action
-        """
-
-        self.logger.debug("Constructing the URL for task deletion")
-        url = "{0}/workspaces/{1}/jobs/{2}/job_tasks/{3}".format(self.base_url, workspace_id, job_id, task_id)
-        url = self._add_token_to_url(url)
-        self.logger.debug("We have constructed the URL for task deletion and is: {0}".format(url))
-        response = self.session.delete(url)
-        self.logger.debug("Received response code {0} with reason {1}...".format(response.status_code, response.reason))
         return response
+
+        # self.session.headers.pop("Content-Type")
+        # self.logger.debug(response.content)
+        # if response.status_code == 200:
+        #     process_id = response.json()['meta']['processId']
+        #
+        #     self.logger.debug("Workflow {0} started with process {1}".format(workflow_id, process_id))
+        #     return process_id
+        # else:
+        #     raise RunFlowFailureException(
+        #         "Run Workflow {0} failed with status code {1}".format(workflow_id, response.status_code))
